Adjusted ICE candidate priority field from ulong to uint to accommodate 4 byte STUN PRIORITY attributes. Resolves #350.

+++ src/net/ICE/RTCIceCandidate.cs
@@ -71,7 +71,7 @@ namespace SIPSorcery.Net
         /// <remarks>
         /// See specification at https://tools.ietf.org/html/rfc8445#section-5.1.2.
         /// </remarks>
-        public ulong priority { get; set; }
+        public uint priority { get; set; }
 
         /// <summary>
         /// The address or hostname for the candidate.
@@ -179,7 +179,7 @@ namespace SIPSorcery.Net
                     candidate.protocol = candidateProtocol;
                 }
 
-                if (ulong.TryParse(candidateFields[3], out var candidatePriority))
+                if (uint.TryParse(candidateFields[3], out var candidatePriority))
                 {
                     candidate.priority = candidatePriority;
                 }
@@ -270,9 +270,9 @@ namespace SIPSorcery.Net
             return (type.GetHashCode() + addressVal + svrVal + protocol.GetHashCode()).ToString();
         }
 
-        private ulong GetPriority()
+        private uint GetPriority()
         {
-            return (ulong)((2 ^ 24) * (126 - type.GetHashCode()) +
+            return (uint)((2 ^ 24) * (126 - type.GetHashCode()) +
                       (2 ^ 8) * (65535) + // TODO: Add some kind of priority to different local IP addresses if needed.
                       (2 ^ 0) * (256 - component.GetHashCode()));
         }
